requirements.txt
requests
bs4
pandas
selenium
openai
python-dotenv